requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
Django==5.0.8
djangorestframework==3.15.2
flower==2.0.1
fonttools==4.53.1
html5lib==1.1
humanize==4.10.0
idna==3.7
kombu==5.4.0
pillow==10.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
redis==5.0.4
requests==2.32.3
setuptools==72.2.0
six==1.16.0
sqlparse==0.5.0
stripe==9.3.0
tinycss2==1.3.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
zopfli==0.2.3
